project/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.1
bootstrap4==0.1.0
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
Django==3.1
django-bootstrap3==14.1.0
django-bootstrap4==2.2.0
django-ckeditor==6.0.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-import-export==2.3.0
django-js-asset==1.2.2
django-widget-tweaks==1.4.8
et-xmlfile==1.0.1
filelock==3.0.12
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.2.0
pyfiglet==0.8.post1
pylint==2.5.3
pytz==2020.1
PyYAML==5.3.1
self==0.0.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tablib==2.0.0
termcolor==1.1.0
toml==0.10.1
virtualenv==20.0.25
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
